Updates for Helm CRD use and ordering:
This uses the Helm native crds/ directory so they get
installed first. Also, tries to add some ordering so that
the stack service type LoadBalancer waits to be installed
until after the LoadBalancer (KubeVIP) is installed. Helm
must be called with `--wait` for this to function correctly
and for the stack service to get the load balancer IP assigned.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>