Update Kubevip version and interface and Boots deployment strategy:
Removing the requirement for an interface for Kubevip allows
Kubevip to auto discover the interface to use. Kubevip does this
by looking for the interface that holds the default route.

Updating Boots' deployment strategy fixes an issue with
single node deployments being unable to upgrade Boots
because of host ports being in use.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>