Update DHCP broadcast interface handling:
Use 127.1.1.1/32 for the DHCP broadcast interface
instead of the load balancer IP. Using the load balancer
IP can cause instability with that address and routing
to Kubernetes services.

Make the DHCP broadcast interface name static. The dynamic
number added to the name Helm to restarts on every Helm deploy.

Add ipvlan support for the DHCP broadcast interface. This allows
deployment where creating and broadcasting a new mac address is
prohibited. Vmware for example.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>